+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ x.x.x ?
 * Fix mappings for Table
 * Added support for AWS Cross-Account in CloudwatchMetricsTarget
 * Added `LokiTarget`
+* Added `color`, `fieldMinMax`, and `thresholdType` option for `GaugePanel`
+* Added `TempoTarget` and `TempoFilter` to target Tempo queries
 
 0.7.1 2024-01-12
 ================